packages/skg-registry/pyproject.toml
[project]
name = "skg-registry"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "skg-protocol>=0.1.0",
  "PyYAML>=6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
